backend/requirements/base.txt
asgiref==3.8.1
black==24.4.0
click==8.1.7
Django==5.0.4
django-cors-headers==4.3.1
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.15.1
flake8==7.0.0
gunicorn==21.2.0
h11==0.14.0
isort==5.13.2
mccabe==0.7.0
mypy==1.9.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pycodestyle==2.11.1
pyflakes==3.2.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
types-pytz==2024.1.0.20240203
types-PyYAML==6.0.12.20240311
typing_extensions==4.11.0
uritemplate==4.1.1
uvicorn==0.29.0
